pyproject.toml
[project]
name = "caller-agent"
version = "0.1.0"
description = "AWS-based automated caller answering agent using Bedrock, Connect, and Lex"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
lambda = [
    "requests>=2.32.0",
    "boto3>=1.35.0",
]
